add test for storage
